% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generate_Data.R
\name{Generate_Data}
\alias{Generate_Data}
\title{Simulate data from a Kalman model}
\usage{
Generate_Data(
  n,
  A,
  C,
  Sigma_Add,
  Sigma_Inn,
  mu_0 = NULL,
  anomaly_loc = integer(0),
  anomaly_type = character(0),
  anomaly_comp = integer(0),
  anomaly_strength = NULL
)
}
\arguments{
\item{n}{A positive integer giving the number of observations desired}

\item{A}{A matrix giving the updates for the hidden states.}

\item{C}{A matrix mapping the hidden states to the observed states.}

\item{Sigma_Add}{A positive definite diagonal matrix giving the additive noise covariance.}

\item{Sigma_Inn}{A positive definite diagonal matrix giving the innovative noise covariance.}

\item{mu_0}{A matrix indicating the mean of the prior for the hidden states. It defaults to a zero-vector.}

\item{anomaly_loc}{A vector of integers giving the locations of anomalies.}

\item{anomaly_type}{A vector of strings, either "Add" or "Inn" indicating whether the anomaly is additive or innovative.}

\item{anomaly_comp}{A vector of integers giving the component affected by the anomalies.}

\item{anomaly_strength}{A vector of numerics givig the strength of the anomalies (in sigmas).}
}
\value{
A list of matrices, each corresponding to an observation.
}
\description{
This function simulates data obeying a Kalman model whilst alloiwng the user to add innovative and additive anomalies.
}
