% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ioaorkf.R
\name{plot.ioaorkf}
\alias{plot.ioaorkf}
\title{Plot method for ioarkf}
\usage{
\method{plot}{ioaorkf}(x, time = NULL, horizon = NULL, subset = NULL, ...)
}
\arguments{
\item{x}{An ioaorkf object.}

\item{time}{A positive integer giving the time at which the output is to be displayed. It defaults to the number of observations}

\item{horizon}{A positive integer giving the smoothing horizon that is to be used. It must be at least equal to the number of rows of the horizonmatric used to obtain the ioaorkf object.}

\item{subset}{A list of integers indicating the components of observations which are to be plotted.}

\item{...}{Other parameters to be passed to plotting methods.}
}
\value{
A ggplot object.
}
\description{
A function to plot the output of an IOAORKF filter. One can specify a time during the run for which the output should be displayed.
}
