% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.rkf.R
\name{plot.rkf}
\alias{plot.rkf}
\title{Plot method for rkf}
\usage{
\method{plot}{rkf}(x, time = NULL, subset = NULL)
}
\arguments{
\item{x}{An ioaorkf object.}

\item{time}{A positive integer giving the time at which the output is to be displayed. It defaults to the number of observations}

\item{subset}{A list of components of observations which are to be plotted.}
}
\value{
A ggplot object.
}
\description{
A function to plot the output of an IOAORKF filter. One can specify a time during the run for which the output should be displayed.
}
