% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.ioaorkf.R
\name{print.ioaorkf}
\alias{print.ioaorkf}
\title{Print method for ioarkf}
\usage{
\method{print}{ioaorkf}(x, time = NULL, horizon = NULL, ...)
}
\arguments{
\item{x}{An ioaorkf object.}

\item{time}{A positive integer giving the time at which the output is to be displayed. It defaults to the number of observations}

\item{horizon}{A positive integer giving the smoothing horizon that is to be used. It must be at least equal to the number of rows of the horizonmatric used to obtain the ioaorkf object.}

\item{...}{Other parameters to be passed to plotting methods.}
}
\value{
A ggplot object.
}
\description{
A function to print the output of an IOAORKF filter. One can specify a time during the run for which the output should be displayed.
}
