% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.rkf.R
\name{print.rkf}
\alias{print.rkf}
\title{Print method for rkf}
\usage{
\method{print}{rkf}(x, time = NULL)
}
\arguments{
\item{x}{An ioaorkf object.}

\item{time}{A positive integer giving the time at which the output is to be displayed. It defaults to the number of observations}
}
\value{
A ggplot object.
}
\description{
A function to print the output of an IOAORKF filter. One can specify a time during the run for which the output should be displayed.
}
