% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.ioaorkf.R, R/summary.rkf.R
\name{summary}
\alias{summary}
\alias{summary.ioaorkf}
\alias{summary.rkf}
\title{Summary}
\usage{
\method{summary}{ioaorkf}(object, time = NULL, horizon = NULL, ...)

\method{summary}{rkf}(object, time = NULL, conf_level = 0.95, ...)
}
\arguments{
\item{object}{An instance of an \code{ioaorkf} or \code{rkf} S3 class.}

\item{time}{A positive integer giving the time at which the output is to be displayed. It defaults to the number of observations.}

\item{horizon}{A positive integer giving the smoothing horizon that is to be used. It must be at least equal to the number of rows of the horizonmatrix used to obtain the ioaorkf object.}

\item{...}{Ignored}

\item{conf_level}{A probability between 0 and 1 giving the confidence level at which the series are to be tested against anomalies. It defaults to 0.95.}
}
\description{
A function to summarise the output produced by \code{\link{AORKF_t}}, \code{\link{AORKF_huber}}, \code{\link{IORKF_huber}}, or \code{\link{IOAORKF}}.
One can specify a time during the run for which the output should be displayed.
}
