% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.rkf.R
\name{summary.rkf}
\alias{summary.rkf}
\title{Summary method for rkf}
\usage{
\method{summary}{rkf}(object, time = NULL, conf_level = 0.95, ...)
}
\arguments{
\item{object}{An rkf object.}

\item{time}{A positive integer giving the time at which the output is to be displayed. It defaults to the number of observations.}

\item{conf_level}{A probability between 0 and 1 giving the confidence level at which the series are to be tested against anomalies. It defaults to 0.95.}

\item{...}{Other parameters to be passed to plotting methods.}
}
\value{
A ggplot object.
}
\description{
A function to summarise the output produced by \code{\link{AORKF_t}}, \code{\link{AORKF_huber}}, or \code{\link{IORKF_huber}}. One can specify a time during the run for which the output should be displayed.
}
